organizing module layout again
