+++ src/client.rs
@@ -1,4 +1,3 @@
-use std::error::Error;
 
 use tokio::runtime::Runtime;
 use log::warn;
