+++ src/notice/error.rs
@@ -1,6 +1,6 @@
 
 use std::error::Error;
-use std::collections::HashMap;
+use super::NoticeBacktrace;
 
 #[derive(Debug, Serialize)]
 pub struct NoticeError {
@@ -23,12 +23,3 @@ impl<'a, E: Error> From<E> for NoticeError {
         }
     }
 }
-
-#[derive(Debug, Serialize)]
-pub struct NoticeBacktrace {
-    pub file: String,
-    pub line: i32,
-    pub function: String,
-    pub code: Option<HashMap<i32, String>>
-}
-
